Added useful_prefetch parameter to prefetchers (#349)
Resolves #344 

This patch adds a parameter to the prefetcher cache_operate hook that
indicates whether the hit block was a prefetch or not. This will help
prefetchers get good feedback on their performance, and it also allows
easier migration from the IPC-1 prefetchers. This is, however, a
breaking change for all existing code in the name of bringing
instruction prefetchers and data prefetchers together.