Added cache fill hook for replacement policies (#434)
This patch adds the ability for replacement policies to observe the
cache miss and cache fill as separate events. This reconciles the
interface to be similar to prefetchers, which have had this feature for
a long time.

If the cache fill hook is declared, the simulator will call the
`update_replacement_state()` hook (or not, if it is absent) at tag check
time, for both hits and misses. The previous behavior delayed the call
to `update_replacement_state()` for misses until fill time.